# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vena - Gestão Inteligente de Compras e Obras
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/request-access:
    post:
      operationId: requestAccess
      tags: [auth]
      summary: Request access to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestInput"
      responses:
        "201":
          description: Access request sent
  /auth/set-password:
    post:
      operationId: setPassword
      tags: [auth]
      summary: Set password for first access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordInput"
      responses:
        "200":
          description: Password set
...
  /auth/users/{id}/approve:
    post:
      operationId: approveUser
      tags: [auth]
      summary: Approve user access request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User approved

  # DASHBOARD
...
  /financial/accounts:
    get:
      operationId: listFinancialAccounts
      tags: [financial]
      summary: List accounts payable and receivable
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [payable, receivable]
        - name: status
          in: query
          schema:
            type: string
        - name: due_in_days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financial accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialAccount"
    post:
      operationId: createFinancialAccount
      tags: [financial]
      summary: Create financial account (payable or receivable)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialAccount"
  /financial/accounts/{id}:
    patch:
      operationId: updateFinancialAccount
      tags: [financial]
      summary: Update financial account (e.g. mark as paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialAccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialAccount"
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials/inventory
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
        - name: low_stock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create material/inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get material by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update material (including stock quantity)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, engineer, foreman]
        department:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending, inactive]
        createdAt:
          type: string
    AccessRequestInput:
      type: object
      required: [name, email, department, position]
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        position:
          type: string
    SetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string

    # DASHBOARD
...
    FinancialAccount:
      type: object
      required: [id, type, description, amount, dueDate, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [payable, receivable]
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, overdue, cancelled]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        createdAt:
          type: string
    FinancialAccountInput:
      type: object
      required: [type, description, amount, dueDate]
      properties:
        type:
          type: string
          enum: [payable, receivable]
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
        projectId:
          type: integer
        supplierId:
          type: integer
        category:
          type: string
    FinancialAccountUpdate:
      type: object
      properties:
        status:
          type: string
        paidAt:
          type: string
        amount:
          type: number
        dueDate:
          type: string
...
    Material:
      type: object
      required: [id, name, category, unit, currentStock, minimumStock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minimumStock:
          type: number
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        lastPurchasePrice:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
    MaterialInput:
      type: object
      required: [name, category, unit, minimumStock]
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minimumStock:
          type: number
        projectId:
          type: integer
        location:
          type: string
        notes:
          type: string
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        currentStock:
          type: number
        minimumStock:
          type: number
        projectId:
          type: integer
        location:
          type: string
        notes:
          type: string
...